Extracting EtwLogActivity class to be available outside NuGetVSActionTelemetryService.
